refactor(server-core): generate the query surface, stop rewriting the document
trapi 2.1.1 landed the three things that were missing, so the post-generation
step is gone and `build:swagger` is a plain `trapi generate`.

The `@DQuerySchema` handler now contributes each marked operation's query
parameters during generation (tada5hi/trapi#907) instead of a later pass
attaching them to the finished document. The shared error response is declared
document-wide (#908), with the 400 staying on the handler because only a read
that decodes a filter can answer one: every other query parameter fails soft,
dropping an unknown key rather than rejecting it. That split is finer than
what it replaces, which derived the same distinction from a predicate in the
rewriting pass.

The console mounts read their constants again (#906). They were spelled inline
because trapi resolved a decorator path only as a literal, which had both
controllers emitting at the document root, colliding on `/login/start` and
`/callback`, and half of their operations dropped.

Verified by diffing the regenerated document against the one the previous
pipeline produced: 229 operations, no operation added or removed, and every
parameter identical except for ordering and the `allowEmptyValue` /
`deprecated` / `examples` / `format` metadata trapi now supplies itself. The
`minimum` on `page[limit]` survives the translation as a validator.

The one regression is that 61 operations lose the `realmId` description, since
a contributed parameter is appended after the derived ones and cannot describe
a variable trapi synthesizes. It is one sentence, and the only place the
document said a realm is addressable by name as well as by id.

Two defects the diff caught before this landed: the three bulk revokes lost
their 400, because the response was keyed on being a collection read rather
than on decoding a filter, and `ErrorResponse` vanished from `components`
leaving every error response a dangling ref, because the lifted
`securityDefinitions` block carried a second `extra` key that silently won.
Both built green.
