+++ apps/server-core/package.json
@@ -135,8 +135,8 @@
     "devDependencies": {
         "@authup/server-test-kit": "^1.0.0-beta.64",
         "@faker-js/faker": "^10.6.0",
-        "@trapi/cli": "^2.1.0",
-        "@trapi/core": "^2.1.0",
+        "@trapi/cli": "^2.1.1",
+        "@trapi/core": "^2.1.1",
         "@types/ldapjs": "^3.0.6",
         "@types/nodemailer": "^8.0.0",
         "@types/qrcode": "^1.5.6",
